@@ -23,8 +23,8 @@
         "@mui/x-date-pickers-pro": "^8.9.2",
         "@savi2w/chess-tcn": "^1.0.0",
         "@types/node": "^24.1.0",
-        "@types/react": "^18.3.12",
-        "@types/react-dom": "^18.3.1",
+        "@types/react": "19.1.9",
+        "@types/react-dom": "19.1.7",
         "async-mutex": "^0.5.0",
         "aws-amplify": "^6.15.4",
         "axios": "^1.11.0",
@@ -36,20 +36,18 @@
         "lodash.debounce": "^4.0.8",
         "luxon": "^3.6.1",
         "modern-screenshot": "^4.6.5",
-        "next": "^14.2.11",
+        "next": "15.4.5",
         "next-navigation-guard": "^0.2.0",
-        "react": "^18.2.0",
+        "react": "19.1.1",
         "react-activity-calendar": "^2.7.13",
         "react-charts": "^3.0.0-beta.57",
-        "react-copy-to-clipboard": "^5.1.0",
         "react-countdown-circle-timer": "^3.2.1",
-        "react-dom": "^18.2.0",
+        "react-dom": "19.1.1",
         "react-ga4": "^2.1.0",
         "react-google-button": "^0.8.0",
         "react-icons": "^5.5.0",
-        "react-joyride": "^2.9.3",
+        "react-joyride-react19-compat": "^2.9.5-react19-compat",
         "react-markdown": "^10.1.0",
-        "react-material-ui-carousel": "^3.4.2",
         "react-minimal-pie-chart": "^9.1.1",
         "react-resizable": "^3.0.5",
         "react-tooltip": "^4.4.3",
@@ -110,6 +108,7 @@
         ]
     },
     "devDependencies": {
+        "@eslint/js": "^9.32.0",
         "@istanbuljs/nyc-config-typescript": "^1.0.2",
         "@jackstenglein/eslint-config-dojo": "file:../eslint-config-dojo",
         "@testing-library/jest-dom": "^6.6.4",
@@ -120,20 +119,17 @@
         "@types/react-copy-to-clipboard": "^5.0.7",
         "@types/react-resizable": "^3.0.8",
         "@types/uuid": "^10.0.0",
-        "@typescript-eslint/eslint-plugin": "^7.12.0",
-        "@typescript-eslint/parser": "^7.12.0",
+        "@typescript-eslint/eslint-plugin": "^8.38.0",
+        "@typescript-eslint/parser": "^8.38.0",
         "@vitejs/plugin-react": "^4.7.0",
         "command-line-args": "^6.0.1",
         "command-line-usage": "^7.0.3",
         "cypress": "^14.5.3",
         "cypress-parallel": "^0.15.0",
         "dotenv": "^17.2.1",
-        "eslint": "^8.57.0",
-        "eslint-config-next": "^15.4.5",
+        "eslint": "^9.32.0",
+        "eslint-config-next": "15.4.5",
         "eslint-plugin-cypress": "^5.1.0",
-        "eslint-plugin-eqeqeq-fix": "^1.0.3",
-        "eslint-plugin-react": "^7.37.5",
-        "eslint-plugin-react-hooks": "^5.2.0",
         "happy-dom": "^17.4.1",
         "json-2-csv": "^5.5.9",
         "nyc": "^17.1.0",
@@ -150,6 +146,8 @@
     "overrides": {
         "d3-interpolate": "^3.0.1",
         "d3-color": "^3.1.0",
-        "d3-scale": "^4.0.2"
+        "d3-scale": "^4.0.2",
+        "@types/react": "19.1.9",
+        "@types/react-dom": "19.1.7"
     }
 }
