@@ -4,16 +4,16 @@ module.exports = {
         '@jackstenglein/eslint-config-dojo/recommended',
         '@jackstenglein/eslint-config-dojo/next',
     ],
-    parser: '@typescript-eslint/parser',
-    parserOptions: {
-        project: true,
-        tsconfigRootDir: __dirname,
-    },
-    settings: {
-        react: {
-            version: 'detect',
-        },
-    },
+    // parser: '@typescript-eslint/parser',
+    // parserOptions: {
+    //     project: true,
+    //     tsconfigRootDir: __dirname,
+    // },
+    // settings: {
+    //     react: {
+    //         version: 'detect',
+    //     },
+    // },
     ignorePatterns: [
         'build/**',
         'node_modules/**',
