@@ -1,5 +1,11 @@
 import { CreateClubPage } from '@/components/clubs/CreateClubPage';
 
-export default function Page({ params: { id } }: { params: { id: string } }) {
+export default async function Page(props: { params: Promise<{ id: string }> }) {
+    const params = await props.params;
+
+    const {
+        id
+    } = params;
+
     return <CreateClubPage id={id} />;
 }
