@@ -81,11 +81,18 @@ function getPlayer(chess: Chess, key: 'White' | 'Black', eloKey: 'WhiteElo' | 'B
     return 'NN';
 }
 
-export default function Page({
-    params: { cohort, id },
-}: {
-    params: { cohort: string; id: string };
-}) {
+export default async function Page(
+    props: {
+        params: Promise<{ cohort: string; id: string }>;
+    }
+) {
+    const params = await props.params;
+
+    const {
+        cohort,
+        id
+    } = params;
+
     return (
         <Suspense>
             <GamePage cohort={cohort} id={id} />
