@@ -5,11 +5,18 @@ export function generateStaticParams() {
     return [{ type: 'followers' }, { type: 'following' }];
 }
 
-export default function Page({
-    params: { username, type },
-}: {
-    params: { username: string; type: 'followers' | 'following' };
-}) {
+export default async function Page(
+    props: {
+        params: Promise<{ username: string; type: 'followers' | 'following' }>;
+    }
+) {
+    const params = await props.params;
+
+    const {
+        username,
+        type
+    } = params;
+
     return (
         <Suspense>
             <FollowersPage username={username} type={type} />
