@@ -1,7 +1,7 @@
 import CloseButton from '@/components/ui/CloseButton';
 import ScoreboardProgress from '@/scoreboard/ScoreboardProgress';
 import { Box, Button, Card, CardActions, CardContent, CardHeader, Typography } from '@mui/material';
-import { TooltipRenderProps } from 'react-joyride';
+import { TooltipRenderProps } from 'react-joyride-react19-compat';
 
 const TutorialTooltip: React.FC<TooltipRenderProps> = ({
     index,
